feat(api): move AlreadyExistingAdminMemberError to src/team
