+++ api/tests/unit/domain/usecases/save-admin-member_test.js
@@ -1,11 +1,11 @@
+import { UserNotFoundError } from '../../../../lib/domain/errors.js';
 import { saveAdminMember } from '../../../../lib/domain/usecases/save-admin-member.js';
 import { PIX_ADMIN } from '../../../../src/authorization/domain/constants.js';
+import { AdminMember } from '../../../../src/shared/domain/models/AdminMember.js';
+import { AlreadyExistingAdminMemberError } from '../../../../src/team/domain/errors.js';
 import { catchErr, domainBuilder, expect, sinon } from '../../../test-helper.js';
 
 const { ROLES } = PIX_ADMIN;
-
-import { AlreadyExistingAdminMemberError, UserNotFoundError } from '../../../../lib/domain/errors.js';
-import { AdminMember } from '../../../../src/shared/domain/models/AdminMember.js';
 
 describe('Unit | UseCase | save-admin-member', function () {
   context('when admin member email is not found', function () {
